Create indexANC.html
Just started :) Will fill in the rest shortly. Wish I could open up my own gaming cafe...This is a H.W. assignment for my first html5 web design class. Not sure how to link a page to the Games page that wont be linked to all the other pages.